Fixed a whole bunch of linting errors

+++ lib/virtual-fs/src/mem_fs/file_opener.rs
@@ -414,7 +414,7 @@ impl crate::FileOpener for FileSystem {
 
                         // Move the cursor to the end if needed.
                         if append {
-                            cursor = file.len() as u64;
+                            cursor = file.len();
                         }
                     }
